# packages: faster should-filter query