--- a/migrations/frontend/1678994673_package_repos_last_checked_at_index/metadata.yaml
+++ b/migrations/frontend/1678994673_package_repos_last_checked_at_index/metadata.yaml
@@ -1,0 +1,2 @@
+name: package_repos_last_checked_at_index
+parents: [1678899992, 1678832491]